Refactor dataset.py: add KDTree for efficient nearest neighbor search
This commit adds a KDTree from the scipy.spatial module to improve the efficiency of nearest neighbor search in the MeshDataset class. The KDTree is used to extract a subset of faces based on their proximity to a randomly generated point within the bounding box of the mesh. This subset is then used to create new vertices and faces for augmentation. Additionally, the maximum number of faces allowed in a mesh has been set to 500.